Move delivery to the package level
